// src/app/State/product-regular/product-types/product-types.ts
export enum ProductActionsType {
  LOAD_PRODUCT = '[Product Page]Load Product',
  LOAD_PRODUCT_SUCCESS = '[Product Page]Load Product Success',
  LOAD_PRODUCT_FAIL = '[Product Page]Load Product Fail',
  ADD_PRODUCT = '[Product Page]Add Product',
  ADD_PRODUCT_SUCCESS = '[Product Page]Add Product Success',
  UPDATE_PRODUCT = '[Product Page]Update Product',
  UPDATE_PRODUCT_SUCCESS = '[Product Page]Update Product Success',
  DELETE_PRODUCT = '[Product Page]Delete Product',
  DELETE_PRODUCT_SUCCESS = '[Product Page]Delete Product Success',
  GET_PRODUCT = '[Product Page]Get Product',
  GET_PRODUCT_SUCCESS = '[Product Page]Get Product Success',
  OPEN_POPUP = '[Product Page]Open Popup',
}